feat(payment): adicionar módulo de pagamentos com integração à reservas e dashboard
- Implementa entidades, DTOs, repositórios, handlers e controller para pagamentos
- Integra criação e exclusão de pagamentos ao fluxo de reservas
- Adiciona cálculo de receita bruta no dashboard e relatórios de reservas
- Permite registro de clientes via endpoint público
- Ajusta tipos de dados para valores monetários em manutenção e pagamento
- Habilita gerenciamento transacional nos handlers

+++ src/main/java/io/github/joabsonlg/sigac_api/reservation/dto/ReservationReportDTO.java
@@ -1,6 +1,8 @@
 package io.github.joabsonlg.sigac_api.reservation.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -10,7 +12,7 @@ public record ReservationReportDTO(
     Long confirmedReservations,
     Long completedReservations,
     Long cancelledReservations,
-    Double totalRevenue,
+    BigDecimal totalRevenue,
     List<ReservationDTO> latestReservations,
     Map<String, Double> statusPercentages
 ) {}
